Move get_num_at_risk_tasks to marathon_tools
